% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticsdata.R
\name{statisticsdata}
\alias{statisticsdata}
\title{Finding for the five number of summary, standard deviation, variance for crude prevalence data}
\usage{
statisticsdata(x)
}
\arguments{
\item{x}{A \code{vector} of dimension 1 that is used to be calculated its summary}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{sd}{Estimated standard deviation of the estimator}
     \item{var}{Estimated variance of the estimator}
}
}
\description{
Compute the five number summary of the data value of the dataset
}
\details{
divide and combine datas accordingly into two groups; crude and age
this is for the crude data
}
\author{
Group 5
}
